chore: Skeleton changes for making messages view draggable
