+++ features/maprealtime/impl/src/main/kotlin/io/element/android/features/maprealtime/impl/MapToolbar.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.element.android.features.location.impl.all.composables
+package io.element.android.features.maprealtime.impl
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -22,11 +22,7 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.TopAppBar
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Phone
-import androidx.compose.material.icons.outlined.Phone
 import androidx.compose.material.icons.outlined.Textsms
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.IconButtonDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -34,14 +30,19 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import io.element.android.compound.tokens.generated.CompoundIcons
+import io.element.android.features.roomcall.api.RoomCallState
 import io.element.android.libraries.designsystem.theme.components.Icon
 import io.element.android.libraries.designsystem.theme.components.IconButton
 import io.element.android.libraries.designsystem.theme.components.Text
-import androidx.compose.material3.IconButton as MuiIconButton
 
-@OptIn(ExperimentalMaterial3Api::class)
 @Composable
-internal fun MapToolbar(onBackPressed: () -> Unit, title: String, onMessagesPressed: () -> Unit, onJoinCallClicked: () -> Unit, isCallOngoing: Boolean) {
+internal fun MapToolbar(
+    onBackPressed: () -> Unit,
+    title: String,
+    onMessagesPressed: () -> Unit,
+    onJoinCallClicked: () -> Unit,
+    roomCallState: RoomCallState,
+) {
 
     TopAppBar(
         title = {
@@ -66,18 +67,10 @@ internal fun MapToolbar(onBackPressed: () -> Unit, title: String, onMessagesPres
             ) {
                 Icon(Icons.Outlined.Textsms, contentDescription = "", tint = Color.White)
             }
-            MuiIconButton(
-                onClick = { onJoinCallClicked() },
-                colors = IconButtonDefaults.iconButtonColors(
-                    containerColor = if (isCallOngoing) Color(0xFF00B548) else Color(0xFF1E1E1E),
-                )
-            ) {
-                if (isCallOngoing) {
-                    Icon(Icons.Filled.Phone, contentDescription = "", tint = Color.White)
-                } else {
-                    Icon(Icons.Outlined.Phone, contentDescription = "", tint = Color.White)
-                }
-            }
+            CallMenuItem(
+                roomCallState = roomCallState,
+                onJoinCallClick = onJoinCallClicked,
+            )
         },
         modifier = Modifier
             .padding(start = 16.dp, top = 10.dp, bottom = 10.dp)
@@ -87,5 +80,3 @@ internal fun MapToolbar(onBackPressed: () -> Unit, title: String, onMessagesPres
         elevation = 8.dp,
     )
 }
-
-
