+++ features/messages/impl/src/main/kotlin/io/element/android/features/messages/impl/MessagesEvents.kt
@@ -16,7 +16,8 @@ sealed interface MessagesEvents {
     data class ToggleReaction(val emoji: String, val eventOrTransactionId: EventOrTransactionId) : MessagesEvents
     data class InviteDialogDismissed(val action: InviteDialogAction) : MessagesEvents
     data object Dismiss : MessagesEvents
-    data object ShowMapClicked : MessagesEvents
+    data object ShowMessages : MessagesEvents
+    data object HideMessages : MessagesEvents
 }
 
 enum class InviteDialogAction {
